+++ provider/datasource_rediscloud_active_active_database.go
@@ -43,6 +43,11 @@ func dataSourceRedisCloudActiveActiveDatabase() *schema.Resource {
 				Type:        schema.TypeFloat,
 				Computed:    true,
 			},
+			"redis_version": {
+				Description: "The redis version of the database",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
 			"support_oss_cluster_api": {
 				Description: "Support Redis open-source (OSS) Cluster API",
 				Type:        schema.TypeBool,
@@ -298,6 +303,11 @@ func dataSourceRedisCloudActiveActiveDatabaseRead(ctx context.Context, d *schema
 	if err := d.Set("dataset_size_in_gb", redis.Float64(*db.CrdbDatabases[0].DatasetSizeInGB)); err != nil {
 		return diag.FromErr(err)
 	}
+
+	if err := d.Set("redis_version", redis.String(*db.RedisVersion)); err != nil {
+		return diag.FromErr(err)
+	}
+
 	if err := d.Set("support_oss_cluster_api", redis.BoolValue(db.SupportOSSClusterAPI)); err != nil {
 		return diag.FromErr(err)
 	}