+++ provider/datasource_rediscloud_pro_database_test.go
@@ -13,12 +13,11 @@ func TestAccDataSourceRedisCloudProDatabase_basic(t *testing.T) {
 
 	testAccRequiresEnvVar(t, "EXECUTE_TESTS")
 
-	name := acctest.RandomWithPrefix(testResourcePrefix)
-	password := acctest.RandString(20)
-	testCloudAccountName := os.Getenv("AWS_TEST_CLOUD_ACCOUNT_NAME")
-
 	const dataSourceById = "data.rediscloud_database.example-by-id"
 	const dataSourceByName = "data.rediscloud_database.example-by-name"
+	password := acctest.RandString(20)
+
+	config := getRedisProDbDatasourceConfig(t, password)
 
 	resource.ParallelTest(t, resource.TestCase{
 		PreCheck:          func() { testAccPreCheck(t); testAccAwsPreExistingCloudAccountPreCheck(t) },
@@ -26,7 +25,7 @@ func TestAccDataSourceRedisCloudProDatabase_basic(t *testing.T) {
 		CheckDestroy:      testAccCheckProSubscriptionDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: fmt.Sprintf(testAccDatasourceRedisCloudProDatabase, testCloudAccountName, name, password),
+				Config: config,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(dataSourceById, "name", "tf-database"),
 					resource.TestCheckResourceAttr(dataSourceById, "protocol", "redis"),
@@ -61,72 +60,22 @@ func TestAccDataSourceRedisCloudProDatabase_basic(t *testing.T) {
 					resource.TestCheckResourceAttrSet(dataSourceByName, "private_endpoint"),
 					resource.TestCheckResourceAttr(dataSourceByName, "enable_default_user", "true"),
 					resource.TestCheckResourceAttr(dataSourceByName, "query_performance_factor", "2x"),
+					resource.TestCheckResourceAttr(dataSourceByName, "redis_version", "7.2"),
 				),
 			},
 		},
 	})
+
 }
 
-const testAccDatasourceRedisCloudProDatabase = `
-data "rediscloud_payment_method" "card" {
-  card_type = "Visa"
-}
-data "rediscloud_cloud_account" "account" {
-  exclude_internal_account = true
-  provider_type = "AWS" 
-  name = "%s"
-}
-resource "rediscloud_subscription" "example" {
-  name = "%s"
-  payment_method_id = data.rediscloud_payment_method.card.id
-  memory_storage = "ram"
-  cloud_provider {
-    provider = data.rediscloud_cloud_account.account.provider_type
-    cloud_account_id = data.rediscloud_cloud_account.account.id
-    region {
-      region = "eu-west-1"
-      networking_deployment_cidr = "10.0.0.0/24"
-      preferred_availability_zones = ["eu-west-1a"]
-    }
-  }
-  creation_plan {
-    memory_limit_in_gb = 1
-    quantity = 1
-    replication=false
-    support_oss_cluster_api=true
-    throughput_measurement_by = "operations-per-second"
-    throughput_measurement_value = 1000
-	query_performance_factor	 = "2x"
-	modules = ["RediSearch"]
-  }
-}
-resource "rediscloud_subscription_database" "example" {
-    subscription_id              = rediscloud_subscription.example.id
-    name                         = "tf-database"
-    protocol                     = "redis"
-    memory_limit_in_gb           = 1
-    data_persistence             = "none"
-    throughput_measurement_by    = "operations-per-second"
-    throughput_measurement_value = 1000
-	password                     = "%s"
-	support_oss_cluster_api	     = true
-	replication				     = false
-    enable_default_user 		 = true
-	query_performance_factor	 = "2x"
-	modules = [
-		{
-			name: "RediSearch"
-		}
-	]
-}
+func getRedisProDbDatasourceConfig(t *testing.T, password string) string {
+	testCloudAccountName := os.Getenv("AWS_TEST_CLOUD_ACCOUNT_NAME")
+	subscriptionName := acctest.RandomWithPrefix(testResourcePrefix)
 
-data "rediscloud_database" "example-by-id" {
-  subscription_id = rediscloud_subscription.example.id
-  db_id = rediscloud_subscription_database.example.db_id
-}
+	content, err := os.ReadFile("./testdata/testAccDatasourceRedisCloudProDatabase.tf")
+	if err != nil {
+		t.Fatalf("failed to read file: %v", err)
+	}
 
-data "rediscloud_database" "example-by-name" {
-  subscription_id = rediscloud_subscription.example.id
-  name = rediscloud_subscription_database.example.name
+	return fmt.Sprintf(string(content), testCloudAccountName, subscriptionName, password)
 }
-`