+++ provider/rediscloud_essentials_subscription_test.go
@@ -26,7 +26,7 @@ func TestAccResourceRedisCloudEssentialsSubscription_Free_CRUDI(t *testing.T) {
 	const resourceName = "rediscloud_essentials_subscription.example"
 	const datasourceName = "data.rediscloud_essentials_subscription.example"
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { testAccPreCheck(t) },
 		ProviderFactories: providerFactories,
 		CheckDestroy:      testAccCheckEssentialsSubscriptionDestroy,
@@ -91,7 +91,7 @@ func TestAccResourceRedisCloudEssentialsSubscription_Paid_CreditCard_CRUDI(t *te
 	const resourceName = "rediscloud_essentials_subscription.example"
 	const datasourceName = "data.rediscloud_essentials_subscription.example"
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { testAccPreCheck(t) },
 		ProviderFactories: providerFactories,
 		CheckDestroy:      testAccCheckEssentialsSubscriptionDestroy,
@@ -157,7 +157,7 @@ func TestAccResourceRedisCloudEssentialsSubscription_Paid_NoPaymentType_CRUDI(t 
 	const resourceName = "rediscloud_essentials_subscription.example"
 	const datasourceName = "data.rediscloud_essentials_subscription.example"
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { testAccPreCheck(t) },
 		ProviderFactories: providerFactories,
 		CheckDestroy:      testAccCheckEssentialsSubscriptionDestroy,
@@ -228,7 +228,7 @@ func TestAccResourceRedisCloudEssentialsSubscription_Paid_Marketplace_CRUDI(t *t
 	const resourceName = "rediscloud_essentials_subscription.example"
 	const datasourceName = "data.rediscloud_essentials_subscription.example"
 
-	resource.ParallelTest(t, resource.TestCase{
+	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { testAccPreCheck(t) },
 		ProviderFactories: providerFactories,
 		CheckDestroy:      testAccCheckEssentialsSubscriptionDestroy,
@@ -310,9 +310,16 @@ data "rediscloud_essentials_plan" "example" {
 	region = "us-east-1"
 }
 
+data "rediscloud_payment_method" "card" {
+	card_type = "Visa"
+	last_four_numbers = "5556"
+}
+
 resource "rediscloud_essentials_subscription" "example" {
 	name = "%s"
 	plan_id = data.rediscloud_essentials_plan.example.id
+	# payment_method = "credit-card"
+	# payment_method_id = data.rediscloud_payment_method.card.id
 }
 
 data "rediscloud_essentials_subscription" "example" {
@@ -341,6 +348,7 @@ data "rediscloud_essentials_subscription" "example" {
 const testAccResourceRedisCloudPaidCreditCardEssentialsSubscription = `
 data "rediscloud_payment_method" "card" {
 	card_type = "Visa"
+	last_four_numbers = "5556"
 }
 
 data "rediscloud_essentials_plan" "example" {
@@ -365,6 +373,7 @@ data "rediscloud_essentials_subscription" "example" {
 const testAccResourceRedisCloudPaidNoPaymentTypeEssentialsSubscription = `
 data "rediscloud_payment_method" "card" {
 	card_type = "Visa"
+	last_four_numbers = "5556"
 }
 
 data "rediscloud_essentials_plan" "example" {