+++ provider/rediscloud_private_service_connect_endpoint_test.go
@@ -62,7 +62,8 @@ func TestAccResourceRedisCloudPrivateServiceConnectEndpoint_CRUDI(t *testing.T) 
 
 const testAccResourceRedisCloudPrivateServiceConnectEndpointProStep1 = `
 data "rediscloud_payment_method" "card" {
-  card_type = "Visa"
+	card_type = "Visa"
+	last_four_numbers = "5556"
 }
 
 resource "rediscloud_subscription" "subscription_resource" {