+++ provider/resource_rediscloud_active_active_subscription_peering_test.go
@@ -112,7 +112,8 @@ func TestAccResourceRedisCloudActiveActiveSubscriptionPeering_gcp(t *testing.T) 
 
 const testAccResourceRedisCloudActiveActiveSubscriptionPeeringAWS = `
 data "rediscloud_payment_method" "card" {
-  card_type = "Visa"
+	card_type = "Visa"
+	last_four_numbers = "5556"
 }
 
 resource "rediscloud_active_active_subscription" "example" {
@@ -151,7 +152,8 @@ resource "rediscloud_active_active_subscription_peering" "test" {
 
 const testAccResourceRedisCloudActiveActiveSubscriptionPeeringGCP = `
 data "rediscloud_payment_method" "card" {
-  card_type = "Visa"
+	card_type = "Visa"
+	last_four_numbers = "5556"
 }
 
 resource "rediscloud_active_active_subscription" "example" {